# backend/gateway/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /comment:
        post:
            tags:
                - CommentService
            description: 创建评论
            operationId: CommentService_CreateComment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.CreateCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.CreateCommentResponse'
        delete:
            tags:
                - CommentService
            description: 删除评论
            operationId: CommentService_RemoveComment
            parameters:
                - name: id
                  in: query
                  description: '@gotags: json:"id,omitempty,string"'
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.RemoveCommentResponse'
    /comment/child:
        post:
            tags:
                - CommentService
            operationId: CommentService_ListChildComment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.ListChildCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.ListChildCommentResponse'
    /comment/video:
        post:
            tags:
                - CommentService
            description: 列出视频的评论
            operationId: CommentService_ListComment4Video
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.ListComment4VideoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.ListComment4VideoResponse'
...
    /favorite:
        post:
            tags:
                - FavoriteService
            operationId: FavoriteService_AddFavorite
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.AddFavoriteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.AddFavoriteResponse'
        delete:
            tags:
                - FavoriteService
            operationId: FavoriteService_RemoveFavorite
            parameters:
                - name: target
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: type
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: id
                  in: query
                  description: '@gotags: json:"id,omitempty,string"'
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.RemoveFavoriteResponse'
    /favorite/video/list:
        post:
            tags:
                - FavoriteService
            operationId: FavoriteService_ListFavoriteVideo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.ListFavoriteVideoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.ListFavoriteVideoResponse'
...
components:
    schemas:
        gateway.v1.AddFavoriteRequest:
            type: object
            properties:
                target:
                    type: integer
                    format: enum
                type:
                    type: integer
                    format: enum
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
        gateway.v1.AddFavoriteResponse:
            type: object
            properties: {}
...
        gateway.v1.Comment:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                parentId:
                    type: string
                    description: '@gotags: json:"parentId,omitempty,string"'
                user:
                    $ref: '#/components/schemas/gateway.v1.CommentUser'
                replyUser:
                    $ref: '#/components/schemas/gateway.v1.CommentUser'
                content:
                    type: string
                date:
                    type: string
                likeCount:
                    type: string
                replyCount:
                    type: string
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/gateway.v1.Comment'
        gateway.v1.CommentUser:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
...
        gateway.v1.CreateCommentRequest:
            type: object
            properties:
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                content:
                    type: string
                parentId:
                    type: string
                    description: '@gotags: json:"parentId,omitempty,string"'
                replyUserId:
                    type: string
                    description: '@gotags: json:"replyUserId,omitempty,string"'
        gateway.v1.CreateCommentResponse:
            type: object
            properties:
                comment:
                    $ref: '#/components/schemas/gateway.v1.Comment'
...
        gateway.v1.ListChildCommentRequest:
            type: object
            properties:
                commentId:
                    type: string
                    description: '@gotags: json:"commentId,omitempty,string"'
                pagination:
                    $ref: '#/components/schemas/gateway.v1.PaginationRequest'
        gateway.v1.ListChildCommentResponse:
            type: object
            properties:
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/gateway.v1.Comment'
                pagination:
                    $ref: '#/components/schemas/gateway.v1.PaginationResponse'
...
        gateway.v1.ListComment4VideoRequest:
            type: object
            properties:
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                pagination:
                    $ref: '#/components/schemas/gateway.v1.PaginationRequest'
        gateway.v1.ListComment4VideoResponse:
            type: object
            properties:
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/gateway.v1.Comment'
                pagination:
                    $ref: '#/components/schemas/gateway.v1.PaginationResponse'
        gateway.v1.ListFavoriteVideoRequest:
            type: object
            properties:
                userId:
                    type: string
                page:
                    type: integer
                    format: int32
                size:
                    type: integer
                    format: int32
        gateway.v1.ListFavoriteVideoResponse:
            type: object
            properties:
                videos:
                    type: array
                    items:
                        $ref: '#/components/schemas/gateway.v1.Video'
                pagination:
                    $ref: '#/components/schemas/gateway.v1.PaginationResponse'
...
        gateway.v1.PaginationRequest:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                size:
                    type: integer
                    format: int32
                sort:
                    type: array
                    items:
                        $ref: '#/components/schemas/gateway.v1.SortField'
        gateway.v1.PaginationResponse:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                count:
                    type: integer
                    format: int32
...
        gateway.v1.RemoveCommentResponse:
            type: object
            properties: {}
        gateway.v1.RemoveFavoriteResponse:
            type: object
            properties: {}
...
        gateway.v1.SortField:
            type: object
            properties:
                field:
                    type: string
                order:
                    type: integer
                    format: enum
...
        gateway.v1.Video:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                author:
                    $ref: '#/components/schemas/gateway.v1.VideoAuthor'
                playUrl:
                    type: string
                coverUrl:
                    type: string
                favoriteCount:
                    type: string
                    description: '@gotags: json:"favoriteCount,omitempty,string"'
                commentCount:
                    type: string
                    description: '@gotags: json:"commentCount,omitempty,string"'
                isFavorite:
                    type: boolean
                title:
                    type: string
                isCollected:
                    type: boolean
                collectedCount:
                    type: string
                    description: '@gotags: json:"collectedCount,omitempty,string"'
        gateway.v1.VideoAuthor:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
...